backend/requirements.txt
apimatic-core==0.2.15
apimatic-core-interfaces==0.1.5
apimatic-requests-client-adapter==0.1.6
asgiref==3.8.1
CacheControl==0.12.14
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.0.7
django-cors-headers==4.4.0
django-import-export==3.2.0
django-tinymce==4.1.0
django-truncate==0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
idna==3.10
jsonpickle==3.0.4
jsonpointer==2.4
msgpack==1.1.0
odfpy==1.4.1
openpyxl==3.1.5
paypal-server-sdk==0.5.2
pillow==11.1.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.1
tablib==3.8.0
typing_extensions==4.12.2
urllib3==2.2.3
xlrd==2.0.1
xlwt==1.3.0
